Turns out double is better for mesureing distance in pixels. Who would have thought??

+++ src/canvas.h
@@ -23,7 +23,7 @@ public:
     int dest_indx = -1;
 
 signals:
-    void done(float cost);
+    void done(double cost);
     void start_set();
 protected:
     void paintEvent(QPaintEvent *event) override;
